+++ packages/polaris-viz/src/components/LineChart/Chart.tsx
@@ -275,6 +275,7 @@ export function Chart({
       >
         {hideXAxis ? null : (
           <XAxis
+            allowLineWrap={xAxisOptions.allowLineWrap}
             chartHeight={height}
             chartX={xAxisBounds.x}
             chartY={xAxisBounds.y}