+++ packages/polaris-viz/src/components/StackedAreaChart/Chart.tsx
@@ -249,6 +249,7 @@ export function Chart({
       >
         {hideXAxis ? null : (
           <XAxis
+            allowLineWrap={xAxisOptions.allowLineWrap}
             chartHeight={height}
             chartX={xAxisBounds.x}
             chartY={xAxisBounds.y}